backend/requirements.txt
alembic==1.15.2
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.5
pytest-flask==1.3.0
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
